Add leaflet and react-leaflet for interactive map
Adds leaflet, react-leaflet, and associated types as
dependencies to enable an interactive map component. The map
component is designed to display service areas.

Add leaflet and react-leaflet for interactive map

Adds leaflet, react-leaflet, and their types as dependencies.
This enables the use of interactive maps in the application. Also
created the interactive map component, and map section in this
commit.